+++ CHANGELOG.md
@@ -2,11 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
-## [Unrealeased]
+## [Unreleased]
 
 ### Changed
 
 - `address` field in `AssetBridgingData` can be either EVM or TVM address
+- Method `SenderFactory.getSender` requires additional parameter `network` when creating wallet wrapper using mnemonic
 
 ## [0.3.6] - 2025-01-15